+++ src/main/java/com/github/gtexpert/blpc/mixins/BLPCMixinLoader.java
@@ -20,7 +20,6 @@ public class BLPCMixinLoader implements ILateMixinLoader {
     public static final Map<String, Boolean> modMixinsConfig = new ImmutableMap.Builder<String, Boolean>()
             .put(Mods.Names.BETTER_QUESTING, true)
             .put(Mods.Names.MODULAR_UI, true)
-            .put(Mods.Names.JOURNEY_MAP, true)
             .build();
 
     @Override